+++ src/codecs/common/string.codec.ts
@@ -1,2 +1,3 @@
 import { string as ioString, Type } from 'io-ts';
+
 export const stringCodec: Type<string> = ioString;
