+++ src/codecs/predict-recurrence.codec.ts
@@ -1,5 +1,6 @@
 import { string as ioString, Type, type as ioType } from 'io-ts';
 import { PredictRecurrence } from '../definitions';
+
 export const predictRecurrenceCodec: Type<PredictRecurrence> = ioType({
   query: ioString
 });
