+++ src/commands/edit-fare/edit-fare.persistence.ts
@@ -2,11 +2,10 @@ import { map as taskEitherMap, TaskEither, tryCatch as taskEitherTryCatch } from
 import type { PoolClient, QueryResult } from 'pg';
 import type { PostgresDb } from '@fastify/postgres';
 import { Errors } from '../../reporter';
-import { Entity, Pending } from '../../definitions';
+import { Entity, Pending, PendingPersistence, ScheduledPersistence } from '../../definitions';
 import { pipe } from 'fp-ts/lib/function';
-import { PendingPersistence, ScheduledPersistence } from '../../persistence/persistence.definitions';
 import { EditedToPersist } from './edit-fare.route';
-import { fromDBtoPendingCandidate, fromDBtoScheduledCandidate } from '../../persistence/persistence-utils';
+import { fromDBtoPendingCandidate, fromDBtoScheduledCandidate } from '../../mappers';
 import { onDatabaseError } from '../../errors';
 
 export const persistEditedFares =
