+++ src/commands/edit-fare/edit-fare.ts
@@ -2,7 +2,7 @@ import { pipe } from 'fp-ts/lib/function';
 import { map as taskEitherMap, TaskEither } from 'fp-ts/TaskEither';
 import { Errors } from '../../reporter';
 import { EditedToPersist, FaresToEdit } from './edit-fare.route';
-import { isOneWay } from '../../domain/utils';
+import { isOneWay } from '../../domain';
 import { toPending } from '../../mappers';
 
 export const editFare = (fareToEdit: TaskEither<Errors, FaresToEdit>): TaskEither<Errors, EditedToPersist> =>
