+++ src/commands/edit-fare/edit-fare.validation.ts
@@ -5,8 +5,7 @@ import { PostgresDb } from '@fastify/postgres';
 import { Entity, Scheduled, ToEdit } from '../../definitions';
 import { type as ioType, Type, union as ioUnion } from 'io-ts';
 import { FaresEdited, FaresToEdit } from './edit-fare.route';
-import { throwEntityNotFoundValidationError } from '../../errors/entity-not-found.validation-error';
-import { $onInfrastructureOrValidationError } from '../../errors/infrastructure-or-validation.error';
+import { $onInfrastructureOrValidationError, throwEntityNotFoundValidationError } from '../../errors';
 import {
   entityCodec,
   externalTypeCheckFor,
