+++ src/definitions/domain.definitions.ts
@@ -1,6 +1,6 @@
 import { Place } from './place.definition';
+
 export type Entity = { id: string };
-export type ReturnToDelete = { idToDelete: string };
 
 export type Driver = {
   identifier: string;
