+++ src/definitions/index.ts
@@ -3,3 +3,4 @@ export * from './journey.definition';
 export * from './location.definition';
 export * from './place.definition';
 export * from './recurrence.definition';
+export * from './persistence.definitions';
