+++ src/errors/index.ts
@@ -1,4 +1,4 @@
-export * from './database.error'
-export * from './entity-not-found.validation-error'
-export * from './external-dependency.error'
-export * from './infrastructure-or-validation.error'
+export * from './database.error';
+export * from './entity-not-found.validation-error';
+export * from './external-dependency.error';
+export * from './infrastructure-or-validation.error';
