+++ src/mappers/index.ts
@@ -1,1 +1,2 @@
 export * from './fare.mapper';
+export * from './persistence.mapper';
