+++ src/queries/scheduled-fares-for-date/scheduled-fares-for-date.route.ts
@@ -1,6 +1,6 @@
 import { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
 import { pipe } from 'fp-ts/function';
-import { fold as taskEitherFold, chain as taskEitherChain } from 'fp-ts/TaskEither';
+import { chain as taskEitherChain, fold as taskEitherFold } from 'fp-ts/TaskEither';
 import { onErroredTask, onSuccessfulTaskWith } from '../../server.utils';
 import { isDateString } from '../../codecs';
 import { Entity, Scheduled } from '../../definitions';
@@ -20,7 +20,7 @@ export const scheduledFaresForTheDateQuery = async (
 ): Promise<void> => {
   server.route({
     method: 'GET',
-    url: '/scheduled-fares-for-date/:date',
+    url: '/scheduled/:date',
     handler: async (req: FareForDateRequest, reply: FastifyReply): Promise<void> => {
       await pipe(
         isDateString.decode(req.params.date),
