+++ src/services/google/distance-matrix/estimate-journey.validation.ts
@@ -8,7 +8,7 @@ import {
   googleMapsDistanceMatrixTransferCodec
 } from './distance-matrix.codec';
 import { JourneyEstimate } from '../../../definitions';
-import { Errors } from '../../../reporter/http-reporter';
+import { Errors } from '../../../reporter';
 import { externalTypeCheckFor, journeyEstimateCodec, journeyEstimateRulesCodec } from '../../../codecs';
 
 export const journeyEstimateValidation = (transfer: unknown): TaskEither<Errors, JourneyEstimate> =>
