+++ src/services/openai/completion/completion.api.ts
@@ -1,9 +1,9 @@
 import { chain as taskEitherChain, TaskEither, tryCatch as taskEitherTryCatch } from 'fp-ts/TaskEither';
-import { Errors } from '../../../reporter/http-reporter';
+import { Errors } from '../../../reporter';
 import { pipe } from 'fp-ts/function';
 import { OpenAICompletionPayload, OpenAICompletionResponseTransfer } from './completion.codec';
 import { completionValidation } from './completion.validation';
-import { onDependencyError } from '../../../errors/external-dependency.error';
+import { onDependencyError } from '../../../errors';
 import axios from 'axios';
 
 export const $openAICompletion =
