+++ src/services/openai/completion/completion.validation.ts
@@ -1,7 +1,7 @@
 import { pipe } from 'fp-ts/function';
 import { TaskEither } from 'fp-ts/lib/TaskEither';
 import { fromEither } from 'fp-ts/TaskEither';
-import { Errors } from '../../../reporter/http-reporter';
+import { Errors } from '../../../reporter';
 import { OpenAICompletionResponseTransfer, openAICompletionTransferCodec } from './completion.codec';
 import { externalTypeCheckFor } from '../../../codecs';
