+++ src/services/openai/predict-recurrence/predict-recurrence.ts
@@ -1,6 +1,6 @@
 import { PredictedRecurrence, PredictRecurrence } from '../../../definitions';
 import { chain as taskEitherChain, TaskEither } from 'fp-ts/TaskEither';
-import { Errors } from '../../../reporter/http-reporter';
+import { Errors } from '../../../reporter';
 import { pipe } from 'fp-ts/function';
 import { $openAICompletion } from '../completion/completion.api';
 import { predictedRecurrenceValidation } from './predicted-recurrence.validation';
